+++ src/alib/py/framework/jupiter/framework/mutation_inovcation/entity_event.py
@@ -24,3 +24,4 @@ class MutationEntityEvent:
     session_index: int
     source: str
     context_str: str
+    data: str
